add control cycle

+++ include/inrof_sim/pursuit.hpp
@@ -93,7 +93,7 @@ private:
     // sample
     int K;
     // control cycle
-    float dt;
+    double dt;
     // max value
     std::vector<double> max_value;
     // goal pose
